requirements.txt
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
Pillow==11.0.0
WTForms==3.2.1
